Introduced markdown_to_json
